Split ssz generic test types from tests
